+++ src/components/content-detail/ContentDetailHeader.jsx
@@ -1,6 +1,6 @@
 /* eslint-disable react/prop-types */
 import styled from 'styled-components';
-import useContentStore from '../hooks/useContentStore';
+import useContentStore from '../../hooks/useContentStore';
 
 const HeaderWrapper = styled.div`
   height: 250px;
