+++ src/components/content-detail/ContentDetailPlatform.test.jsx
@@ -3,7 +3,7 @@ import {
   render, screen, waitFor,
 } from '@testing-library/react';
 import { ThemeProvider } from 'styled-components';
-import defaultTheme from '../styles/defaultTheme';
+import defaultTheme from '../../styles/defaultTheme';
 import ContentDetailPlatform from './ContentDetailPlatform';
 
 beforeEach(() => {
