+++ src/components/content-detail/ContentItem.jsx
