+++ src/components/login-signup/LoginForm.test.jsx
@@ -2,7 +2,7 @@ import {
   fireEvent, render, screen, waitFor,
 } from '@testing-library/react';
 import { ThemeProvider } from 'styled-components';
-import defaultTheme from '../styles/defaultTheme';
+import defaultTheme from '../../styles/defaultTheme';
 import LoginForm from './LoginForm';
 
 const navigate = jest.fn();
