+++ src/pages/ContentDetailPage.jsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { useParams } from 'react-router-dom';
+import ContentDetail from '../components/content-detail/ContentDetail';
 import useContentStore from '../hooks/useContentStore';
-import ContentDetail from '../components/ContentDetail';
 
 export default function ContentsDetailPage() {
   const contentStore = useContentStore();
