+++ src/pages/LoginPage.jsx
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import useUserStore from '../hooks/useUserStore';
 import useLoginFormStore from '../hooks/useLoginFormStore';
-import LoginForm from '../components/LoginForm';
+import LoginForm from '../components/login-signup/LoginForm';
 
 export default function LoginPage() {
   const userStore = useUserStore();
